+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Transition.hs
@@ -41,6 +41,8 @@ instance Crypto c => EraTransition (AlonzoEra c) where
 
   mkTransitionConfig = AlonzoTransitionConfig
 
+  injectIntoTestState = registerInitialFundsThenStaking
+
   tcPreviousEraConfigL =
     lens atcMaryTransitionConfig (\atc pc -> atc {atcMaryTransitionConfig = pc})