+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Transition.hs
@@ -21,19 +21,12 @@ module Cardano.Ledger.Api.Transition (
   tcInitialFundsL,
   tcInitialStakingL,
   createInitialState,
-  registerInitialFunds,
-  registerInitialStaking,
-  registerInitialDReps,
-  registerDelegs,
+  injectIntoTestState,
 ) where
 
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import Cardano.Ledger.Api.Era (LatestKnownEra)
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
-import Cardano.Ledger.Conway.Transition (
-  registerDelegs,
-  registerInitialDReps,
- )
 import Cardano.Ledger.Crypto
 import Cardano.Ledger.Shelley.Genesis (ShelleyGenesis (..))
 import Cardano.Ledger.Shelley.Transition (
@@ -41,8 +34,6 @@ import Cardano.Ledger.Shelley.Transition (
   TransitionConfig,
   createInitialState,
   mkShelleyTransitionConfig,
-  registerInitialFunds,
-  registerInitialStaking,
   tcInitialFundsL,
   tcInitialStakingL,
  )